requirements.txt
flask
werkzeug
torch
torchvision
pillow
numpy
